requirements.txt
# Tazkiyah - Quran RAG Platform
# Install with: pip install -r requirements.txt

# ============================================
# Data Collection
# ============================================
requests                    # HTTP client for Quran API
tqdm                        # Progress bars

# ============================================
# RAG Pipeline (LangChain + Ollama)
# ============================================
langchain                   # Core LangChain framework
langchain-ollama            # Ollama integration (embeddings + LLM)
langchain-chroma            # ChromaDB vector store integration
langchain-core              # Core abstractions (prompts, documents, runnables)

# ============================================
# RAG v2 Additions
# ============================================
python-dotenv               # Load .env configuration
langsmith                   # LangSmith tracing & observability

# ============================================
# CLI and UI
# ============================================
click                       # Command-line interface
rich                        # Beautiful terminal output
gradio                      # Web UI for chat interface

# ============================================
# Utilities
# ============================================
numpy                       # Numerical operations
beautifulsoup4              # HTML parsing for chunk preparation
